Adds trigonometric functions sin, cos, and tan (#6)
* Adds trigonetric functions sin, cos, and tan

These are computed with their Maclaurin expansion across the first
quadrant of the unit circle.  Everything beyond that is calculated by
exploiting the symmetry of the sine curve.

Adjust MACLAURIN_SERIES_TERMS to increase/decrease the accuracy and
speed.

Writing tests for these would be silly due to their inaccurate nature.

* Added a link to trig.py to README.md

* Added tests for trig functions

The tests pass if the output is within 1e-6 of the expected result.